fix duplication of events
